+++ actionpack/CHANGELOG.md
@@ -1,3 +1,8 @@
+*   Fix non-GET requests not updating cookies in `ActionController::TestCase`.
+
+    *Jon Moss*, *Hartley McGuire*
+
+
 ## Rails 7.2.1 (August 22, 2024) ##
 
 *   Fix `Request#raw_post` raising `NoMethodError` when `rack.input` is `nil`.